fix: status do agendamento começa com agendado marcado por padrão

+++ vSpring/autoagenda/src/main/java/br/com/autoagenda/autoagenda/repositorios/ProdutoRepository.java
@@ -14,5 +14,5 @@ public interface ProdutoRepository extends CrudRepository<Produto, Integer>{
 	long countByEstoqueAtualLessThan(int nivelCondicao);
 	
 	 @Query("SELECT SUM(p.precoCusto) FROM Produto p")
-	 Float sumTotalPrecoCusto(); 
+	 Double sumTotalPrecoCusto(); 
 }
